added 12

+++ Tasks/BasicOfSoftwareCodeDevelopment/Cycles_4.java
@@ -8,7 +8,7 @@ public class Cycles_4 {
     public static void main(String[] args) {
         BigInteger s = BigInteger.valueOf(1);
         for (int i = 1; i <= 200; i++) {
-            s = s.multiply(BigInteger.valueOf((long) i * i));
+            s = s.multiply(BigInteger.valueOf( i * i));
         }
         System.out.println(s);
